v4.0.29/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
websockets>=11.0
aiofiles>=23.0.0
pyyaml>=6.0
